drop compute from analyze function names
