+++ geoglows/__init__.py
@@ -1,11 +1,10 @@
 import geoglows.bias
 import geoglows.plots
 import geoglows.data
-import geoglows.analysis
-import geoglows.examples
+import geoglows.analyze
 import geoglows.streams
 
-__all__ = ['bias', 'plots', 'data', 'analysis', 'examples', 'streams']
+__all__ = ['bias', 'plots', 'data', 'analyze', 'streams']
 __version__ = '1.0.0'
 __author__ = 'Riley Hales'
 __license__ = 'BSD 3-Clause Clear License'
